[#1508] Organize config schema (separate files)
