+++ src/schemas/config/plans.ts
@@ -1,0 +1,25 @@
+// src/schemas/config/plans.ts
+
+// import { z } from 'zod/v4';
+// import { nullableString } from './shared/primitives';
+
+// plans:
+//   anonymous:
+//     name: 'Anonymous'
+//     ttl: 604800 # 7.days
+//     size: 100_000
+//     api: true
+//   basic:
+//     name: 'Basic Plan'
+//     ttl: 1209600 # 14.days
+//     size: 1_000_000
+//     api: true
+//     email: true
+//     custom_domains: false
+//   identity:
+//     name: 'Identity'
+//     ttl: 2592000.0 # 30.days
+//     size: 10_000_000
+//     api: true
+//     email: true
+//     custom_domains: true
