+++ src/schemas/config/site.ts
@@ -1,0 +1,37 @@
+// src/schemas/config/site.ts
+
+import { z } from 'zod/v4';
+import { nullableString } from './shared/primitives';
+
+const siteAuthenticationSchema = z.object({
+  enabled: z.boolean().default(false),
+  colonels: z.array(z.string()).default([]),
+});
+
+const siteAuthenticitySchema = z.object({
+  enabled: z.boolean().default(false),
+  type: z.string().optional(),
+  secret_key: z.string().optional(),
+});
+
+const siteMiddlewareSchema = z.object({
+  static_files: z.boolean().default(true),
+  utf8_sanitizer: z.boolean().default(true),
+  http_origin: z.boolean().optional(),
+  escaped_params: z.boolean().optional(),
+  xss_header: z.boolean().optional(),
+  frame_options: z.boolean().optional(),
+  path_traversal: z.boolean().optional(),
+  cookie_tossing: z.boolean().optional(),
+  ip_spoofing: z.boolean().optional(),
+  strict_transport: z.boolean().optional(),
+});
+
+const sectionSiteSchema = z.object({
+  host: z.string().default('localhost:3000'),
+  ssl: z.boolean().default(false),
+  secret: z.string().default('CHANGEME'),
+  authentication: siteAuthenticationSchema,
+  authenticity: siteAuthenticitySchema,
+  middleware: siteMiddlewareSchema,
+});
